Rename resources to resource
