+++ .gitignore
@@ -17,3 +17,4 @@ local_strings.py
 coldfront.db
 *.code-workspace
 .vscode
+db.json
