+++ coldfront/core/resource/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ResourceConfig(AppConfig):
+    name = 'resource'
