+++ coldfront/core/resource/views.py
