+++ coldfront/plugins/slurm/management/commands/slurm_dump.py
@@ -3,7 +3,7 @@ import os
 
 from django.core.management.base import BaseCommand, CommandError
 
-from coldfront.core.resources.models import ResourceAttribute
+from coldfront.core.resource.models import ResourceAttribute
 from coldfront.plugins.slurm.utils import SLURM_CLUSTER_ATTRIBUTE_NAME
 from coldfront.plugins.slurm.associations import SlurmCluster
